requirements.txt
selenium
pytest
pytest-selenium
requests
PyYAML